+++ src/Source.Nuke/Tasks/Materials.cs
@@ -19,7 +19,7 @@ namespace Nuke.Common.Tools.Source
             public string DepotDirectory { get; set; }
             public string Format { get; set; } = "tga";
             public string GameName { get; set; }
-            public int AppId { get; set; }
+            public long AppId { get; set; }
         }
 
         // ReSharper disable once CognitiveComplexity
